requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
numpy
scikit-image
python-multipart
python-dotenv
httpx
aiogram